arch: move pg to lib
