+++ api/src/pdc/providers/carpool/repositories/CarpoolRequestRepository.ts
@@ -3,12 +3,12 @@ import {
   PoolClient,
   PostgresConnection,
 } from "@/ilos/connection-postgres/index.ts";
+import sql, { join, raw } from "@/lib/pg/sql.ts";
+import { DatabaseException } from "../exceptions/DatabaseException.ts";
 import {
   InsertableCarpoolRequest,
   WrittenCarpoolRequest,
 } from "../interfaces/index.ts";
-import sql, { join, raw } from "../helpers/sql.ts";
-import { DatabaseException } from "../exceptions/DatabaseException.ts";
 
 @provider()
 export class CarpoolRequestRepository {
